+++ services/vehicle_configurations.go
@@ -28,7 +28,7 @@ type vehicleDAO interface {
 	CountViolations(rs app.RequestScope, deviceid string, reason string) (int, error)
 	GetViolationsByDeviceID(rs app.RequestScope, deviceid string, reason string, offset, limit int) ([]models.DeviceData, error)
 	GetOverspeedByDeviceID(rs app.RequestScope, deviceid string, offset, limit int) ([]models.DeviceData, error)
-	DeleteOverspeedsByDeviceID(rs app.RequestScope, id uint32) (int, error)
+	DeleteOverspeedsByDeviceID(rs app.RequestScope, id uint64) (int, error)
 	SearchVehicles(rs app.RequestScope, searchterm string, offset, limit int, qtype string) ([]models.SearchDetails, error)
 	CountSearches(rs app.RequestScope, searchterm, qtype string) (int, error)
 	UpdatDeviceConfigurationStatus(rs app.RequestScope, deviceid int64, vehicleid uint32) error
@@ -129,7 +129,7 @@ func (s *VehicleService) GetOverspeedByDeviceID(rs app.RequestScope, deviceid st
 }
 
 // DeleteOverspeedsByDeviceID ...
-func (s *VehicleService) DeleteOverspeedsByDeviceID(rs app.RequestScope, id uint32) (int, error) {
+func (s *VehicleService) DeleteOverspeedsByDeviceID(rs app.RequestScope, id uint64) (int, error) {
 	return s.dao.DeleteOverspeedsByDeviceID(rs, id)
 }
